@@ -13,7 +13,7 @@ const posts = [
     POST /api/posts
     { title, body }
 */
-exports.write = (ctx) => {
+export const write = (ctx) => {
     // REST API의 Request Body는 ctx.request.body에서 조회 가능
     const { title, body } = ctx.request.body;
     postId += 1;
@@ -25,14 +25,14 @@ exports.write = (ctx) => {
 /*  포스트 목록 조회
     GET /api/posts
  */
-exports.list = (ctx) => {
+export const list = (ctx) => {
     ctx.body = posts;
 };
 
 /*  특정 포스트 조회
     GET /api/posts/:id
 */
-exports.read = (ctx) => {
+export const read = (ctx) => {
     const { id } = ctx.params;
     const post = posts.find((p) => p.id.toString() === id);
     // 포스트가 없으면 오류를 반환함
@@ -49,7 +49,7 @@ exports.read = (ctx) => {
 /*  특정 포스트 제거
     DELETE /api/posts/:id
 */
-exports.remove = (ctx) => {
+export const remove = (ctx) => {
     const { id } = ctx.params;
     const index = posts.findIndex((p) => p.id.toString() === id);
     // 포스트가 없으면 오류를 반환함
@@ -69,7 +69,7 @@ exports.remove = (ctx) => {
     PUT /api/posts/:id
     { title, body }
 */
-exports.replace = (ctx) => {
+export const replace = (ctx) => {
     const { id } = ctx.params;
     const index = posts.findIndex((p) => p.id.toString() === id);
     if (index === -1) {
@@ -90,7 +90,7 @@ exports.replace = (ctx) => {
     PATCH /api/posts/:id
     { title, body }
 */
-exports.update = (ctx) => {
+export const update = (ctx) => {
     const { id } = ctx.params;
     const index = posts.findIndex((p) => p.id.toString() === id);
     if (index === -1) {
