@@ -1,21 +1,5 @@
-const Koa = require('koa');
-const Router = require('koa-router');
-const bodyParser = require('koa-bodyparser');
+// 여기에서만 비활성
+/* eslint-disable no-global-assign */
 
-const api = require('./api');
-
-const app = new Koa();
-const router = new Router();
-
-// 라우터 설정
-router.use('/api', api.routes()); // api 라우트 적용
-
-// 라우터 적용 전에 bodyParser 적용
-app.use(bodyParser());
-
-// app 인스턴스에 라우터 적용
-app.use(router.routes()).use(router.allowedMethods());
-
-app.listen(4000, () => {
-    console.log('Listening to port 4000');
-});
+require = require('esm')(module /*, options*/);
+module.exports = require('./main.js');
